\name{EMundirStabCov-package}
\alias{EMundirStabCov-package}
\alias{EMundirStabCov}
\docType{package}
\title{
\packageTitle{EMundirStabCov}
}
\description{
\packageDescription{EMundirStabCov}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMundirStabCov}
\packageIndices{EMundirStabCov}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EMundirStabCov}

Maintainer: \packageMaintainer{EMundirStabCov}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
